// src/lib/api/generated/model/emailContact.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */
import type { EmailContactStatusEnum } from './emailContactStatusEnum';

export interface EmailContact {
  readonly id: string;
  readonly email_list: string;
  readonly tenant: string;
  /** @maxLength 255 */
  name: string;
  email: string;
  /** @maxLength 30 */
  phone?: string;
  custom_fields?: unknown;
  consent?: boolean;
  status?: EmailContactStatusEnum;
  readonly created_at: string;
  readonly updated_at: string;
}