// src/lib/api/generated/model/emailContactStatusEnum.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `ACTIVE` - Active
* `UNSUBSCRIBED` - Unsubscribed
* `BOUNCED` - Bounced
 */
export type EmailContactStatusEnum = typeof EmailContactStatusEnum[keyof typeof EmailContactStatusEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EmailContactStatusEnum = {
  ACTIVE: 'ACTIVE',
  UNSUBSCRIBED: 'UNSUBSCRIBED',
  BOUNCED: 'BOUNCED',
} as const;